test(torghut): cover main api refactor wiring
